Moved ExportMonikerWindow out of the "ToolWindows" namespace and into the "Windows" namespace.
