+++ src/KnownMonikersExplorerShared/Windows/ExportMonikerWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="KnownMonikersExplorer.ToolWindows.ExportMonikerWindow"
+﻿<Window x:Class="KnownMonikersExplorer.Windows.ExportMonikerWindow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
